pyproject.toml
[tool.poetry]
name = "logitrackapi"
version = "0.1.0"
description = "API de logística com FastAPI"
authors = ["Jonas Augusto <jonasaugustosantosoliveira@gmail.com>"]
packages = [
    { include = "infrastructure", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.47"}
psycopg = {extras = ["binary"], version = "^3.3.3"}
redis = "^7.2.0"
pydantic-settings = "^2.13.1"
email-validator = "^2.1.0"
fastapi-swagger-dark = "^0.0.9"
alembic = "^1.18.4"
python-dotenv = "^1.2.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
black = "^26.1.0"
isort = "^8.0.0"
httpx = "^0.28.1"
ruff = "^0.15.2"
mypy = "^1.19.1"
pre-commit = "^4.5.1"
pytest-cov = "^7.0.0"
pytest-asyncio = "^1.3.0"

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-ra --cov=src --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]



[tool.coverage.run]
source = ["src"]
omit = [
    "*/__init__.py",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_defs = false